Avoid long pauses on serial timeout on boot on physical hardware.
The Pintos loader writes status to the serial port as well as the
VGA console, but this doesn't work out so well on at least some
real hardware when nothing is connected to the serial port: each
character can take a second or so to display because the BIOS
waits for a serial timeout each time.  So, on the first serial
timeout, replace the "int $0x14" call by a pair of no-ops.

The rest of the changes are part of the struggle not to exceed
the maximum loader size.